Move package ownership checking into WriteAccess
